Update submodules
- Fixed breaking changes
- Fixed some inconsistencies in error messages

+++ src/luagd_permissions.cpp
@@ -14,7 +14,7 @@ void luaGD_checkpermissions(lua_State *L, const char *name, ThreadPermissions pe
         luaL_error(
             L,
             "!!! THREAD PERMISSION VIOLATION: attempted to access %s. needed permissions: %d, got: %d !!!",
-            name, permissions, udata->permissions.operator uint32_t()
+            name, permissions, udata->permissions.operator int64_t()
         );
     }
 }
